refactor: 이모지 스타일 config 파일 추가 (#39)
